Refactor: Ingredients page refactored to IngredientInfo, RelatedIngredient #4
