+++ src/components/Button/StandardButton.jsx
@@ -12,6 +12,7 @@ const StyleButton = styled.button`
   cursor: pointer;
 `;
 
+// onClick이나 추가적인 요소들을 넣을려면 ...rest는 필수
 export default function StandardButton({ width, height, accent, children, ...rest }) {
   return (
     <StyleButton width={width} height={height} data-accent={accent} {...rest}>
